to base that sql optimized
