+++ base/src/main/assets/notes.sql
